hotfix adn add transaction app

+++ financialsystem/payment/models.py
@@ -30,7 +30,7 @@ class Payment(models.Model):
     commission_to = models.ForeignKey(Comission, on_delete=models.SET_NULL, null=True, blank=True)
     payment_mov = models.ForeignKey(Movement, on_delete=models.CASCADE, null=True, blank=True)
     payment_method = models.CharField(max_length=20,choices=MONEY_TYPE, help_text="Metodo de Pago")
-    detail = models.CharField(max_length=50, null=True, blank=True)
+    detail = models.CharField(max_length=150, null=True, blank=True)
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
@@ -69,6 +69,7 @@ def comission_create_inst(instance, *args, **kwargs):
     if instance.installment:
         detail = f'COBRO CUOTA {instance.installment.installment_number} - CLIENTE {instance.installment.credit.client} '
     else:
+        print(instance.installment_ref.refinancing)
         detail = f'COBRO CUOTA {instance.installment_ref.installment_number} - CLIENTE {instance.installment_ref.refinancing.installment_ref.last().credit.client} '
 
     instance.commission_to = Comission.objects.create(
@@ -86,8 +87,11 @@ def delete_mov_commission(instance, *args, **kwargs):
     """
     Elimina el movimiento y comision luego antes de borrar el pago.
     """
-    instance.commission_to.delete()
-    instance.payment_mov.delete()
+    if instance.payment_mov:
+        instance.payment_mov.delete()
+
+    if instance.commission_to:
+        instance.commission_to.delete()
 
 pre_save.connect(up_installment, sender = Payment)
 pre_delete.connect(delete_mov_commission, sender = Payment)
